Refactor CLI into modular components
